# Benchmark robot state