### Description

This change simplifies memory management in `RobotState`, at the expense slower copy/construct operations.
Instead of having a `void *memory_` chunk of memory with crafted pointers pointing inside, this change updates the pointer variables to be what they really are: `std::vector`. 
Construction is more expensive because now it has to do multiple allocation/initialization, instead of a single one. On the positive side, the code is more readable, maintainable and less error-prone. This change actually fixes at least a couple of memory issues with the existing code.

On the performance check, these are the `RobotState` benchmark results for the related test cases:
```
With this PR:
---------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations
---------------------------------------------------------------------------------
robotStateConstruct/100                     0.020 ms        0.020 ms        35180
robotStateConstruct/1000                    0.197 ms        0.197 ms         3650
robotStateConstruct/10000                    1.95 ms         1.95 ms          348
robotStateCopy/100                          0.024 ms        0.024 ms        32654
robotStateCopy/1000                         0.197 ms        0.197 ms         3160
robotStateCopy/10000                         2.09 ms         2.09 ms          318
robotStateUpdate/10                         0.029 ms        0.029 ms        22723
robotStateUpdate/100                        0.282 ms        0.282 ms         2464
robotStateUpdate/1000                        2.72 ms         2.72 ms          252
robotStateForwardKinematics/10              0.025 ms        0.025 ms        28880
robotStateForwardKinematics/100             0.245 ms        0.245 ms         2902
robotStateForwardKinematics/1000             2.48 ms         2.48 ms          285

Without this change (i.e. at HEAD):
---------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations
---------------------------------------------------------------------------------
robotStateConstruct/100                     0.006 ms        0.006 ms       121103
robotStateConstruct/1000                    0.057 ms        0.057 ms        12037
robotStateConstruct/10000                   0.584 ms        0.584 ms         1240
robotStateCopy/100                          0.007 ms        0.007 ms       107305
robotStateCopy/1000                         0.062 ms        0.062 ms        11115
robotStateCopy/10000                        0.628 ms        0.628 ms         1133
robotStateUpdate/10                         0.028 ms        0.028 ms        25451
robotStateUpdate/100                        0.277 ms        0.277 ms         2450
robotStateUpdate/1000                        2.78 ms         2.78 ms          254
robotStateForwardKinematics/10              0.024 ms        0.024 ms        29117
robotStateForwardKinematics/100             0.242 ms        0.242 ms         2847
robotStateForwardKinematics/1000             2.41 ms         2.41 ms          288

```

So, state updates and FK are the same. Copy and construction are around 4x slower, e.g. copying 10.000 states changes from 0.6ms to ~2ms.

Regarding trajectories:
```
With RS Change:
-----------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations
-----------------------------------------------------------------------
robotTrajectoryCreate/10          0.008 ms        0.008 ms        91556
robotTrajectoryCreate/100         0.089 ms        0.089 ms         7889
robotTrajectoryCreate/1000         1.11 ms         1.11 ms          620
robotTrajectoryCreate/10000        22.2 ms         22.2 ms           32
robotTrajectoryCreate/100000        244 ms          244 ms            3
robotTrajectoryTiming/10          0.022 ms        0.022 ms        29042
robotTrajectoryTiming/100         0.070 ms        0.070 ms         9982
robotTrajectoryTiming/1000        0.322 ms        0.322 ms         2161
robotTrajectoryTiming/10000        1.22 ms         1.22 ms          440
robotTrajectoryTiming/20000        1123 ms         1123 ms            1

Without this change (i.e. at HEAD):
-----------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations
-----------------------------------------------------------------------
robotTrajectoryCreate/10          0.005 ms        0.005 ms       100000
robotTrajectoryCreate/100         0.050 ms        0.050 ms        13940
robotTrajectoryCreate/1000         1.10 ms         1.10 ms          609
robotTrajectoryCreate/10000        16.1 ms         16.1 ms           43
robotTrajectoryCreate/100000        190 ms          190 ms            4
robotTrajectoryTiming/10          0.020 ms        0.020 ms        34208
robotTrajectoryTiming/100         0.063 ms        0.063 ms        10910
robotTrajectoryTiming/1000        0.298 ms        0.298 ms         2342
robotTrajectoryTiming/10000       0.975 ms        0.975 ms          554
robotTrajectoryTiming/20000        1057 ms         1057 ms            1
```

Creating a trajectory with 100.000 states goes from 190ms to 244ms.

### Checklist
- [ ] **Required by CI**: Code is auto formatted using [clang-format](http://moveit.ros.org/documentation/contributing/code)
- [ ] Extend the tutorials / documentation [reference](http://moveit.ros.org/documentation/contributing/)
- [ ] Document API changes relevant to the user in the [MIGRATION.md](https://github.com/ros-planning/moveit2/blob/main/MIGRATION.md) notes
- [ ] Create tests, which fail without this PR [reference](https://moveit.picknik.ai/humble/doc/examples/tests/tests_tutorial.html)
- [ ] Include a screenshot if changing a GUI
- [ ] While waiting for someone to review your request, please help review [another open pull request](https://github.com/ros-planning/moveit2/pulls) to support the maintainers

[//]: # "You can expect a response from a maintainer within 7 days. If you haven't heard anything by then, feel free to ping the thread. Thank you!"