@marioprats 2023-12-18
Hi @rhaschke, thanks for following up on this.
I understand the motivation behind the previous memory allocation approach, but it came with added complexity and increased risk for bugs, for a little gain in performance (~2ms over creation of 10k states) that we can probably recover in different ways (e.g. not always allocating velocities and accelerations or some of the transforms).

One example of an issue is the one referenced above in MTC (https://github.com/ros-planning/moveit_task_constructor/issues/518). That's because of an out-of-bounds memory access in `RobotState`.

Here's another example, part of this PR:

```
memset(velocity_, 0, sizeof(double) * 2 * robot_model_->getVariableCount());
```

That was setting `velocity_` to zero, but incorrectly doing it for double the size as I understand.

This PR also fixed an issue in `updateLinkTransformsInternal` that we found when running the tests and discovering out-of-bounds access.

The previous referenced issue in MTC made me realize there's a few other functions in `RobotState` that can trigger out-of-bounds access. I'll prepare a change to fix those.