@rhaschke 2023-12-18
The added complexity was nicely hidden in the init procedure. A user of RobotState doesn't notice.
You only measured performance only for instantiation and FK. What about cache usage in heavy planning methods with thousands of states randomly accessed?
So far you named issues in edge cases only (robots having zero variables). I'm curious to see the underlying reason for https://github.com/ros-planning/moveit_task_constructor/issues/518!
The other example you give, initializing velocity_ to zero, simply zeroes both, velocities and accelerations in one call. No issue there.